fix: remove left sidebar border
This commit removes the left border of sidebar container, because neither of
main menu, status bar or the right edge of the editor control specify a border
at main window edges.

As e.g. Gnome doesn't render any window border, it was pretty obvious and off.

Windows 11 renders thin 1px borders around windows, which caused the left
sidebar border to double up with it - also looks off a bit, as it seems to
belong to main window border.